+++ springcore-20190324/src/main/java/springcore01/homework/md/Springmini篇总结.md
@@ -1,0 +1,10 @@
+1.IOC
+全称控制反转，意思是由将原本由程序员控制的bean的创建、管理的权限转交给容器来做。
+通过扫描设定路径下的类，筛选出需要容器管理的bean，通过反射进行实例化，并保存在IOC容器中
+
+2.DI
+全称依赖注入，spring中contrller和service的属性都没有进行赋值，由容器来进行赋值的过程便是DI，通过反射获取bean的属性，并进行赋值和注入
+
+3.MVC
+全称模型、视图、控制
+通过url在容器handlermapping查找到对应的method实例，通过反射调用方法
